+++ tests/Unit/Logs/TestHook/PhpWarningTest.php
@@ -13,6 +13,14 @@ use PHPUnit\Event\Test\PhpWarningTriggered;
  */
 class PhpWarningTest extends AbstractTestHookTestCase
 {
+    public function testIgnoredByBaseline(): void
+    {
+        $this->createSubscriber()->notify($this->createEvent(ignoredByBaseline: true));
+
+        $logData = $this->getDeserializedLogData();
+        $this->assertTrue($logData->isIgnoredByBaseline(), 'ignoredByBaseline flag should be true');
+    }
+
     protected function createSubscriber(): PhpWarning
     {
         return new PhpWarning();
@@ -23,7 +31,7 @@ class PhpWarningTest extends AbstractTestHookTestCase
         return LogStatus::WarningTriggered;
     }
 
-    protected function createEvent(): PhpWarningTriggered
+    protected function createEvent(bool $ignoredByBaseline = false): PhpWarningTriggered
     {
         return new PhpWarningTriggered(
             $this->createTelemetryInfo(),
@@ -32,7 +40,7 @@ class PhpWarningTest extends AbstractTestHookTestCase
             __FILE__,
             1,
             false,
-            false,
+            $ignoredByBaseline,
         );
     }