+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file. For previou
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
-* ...
+* Suppress output for test issues that are ignored [#381](https://github.com/facile-it/paraunit/pull/381)
 
 ## [2.9.0] - 2026-04-20
 * Add coverage cache warmup [#379](https://github.com/facile-it/paraunit/pull/379)